+++ src/gui/mainMenu.js
@@ -1,58 +1,39 @@
-import {GamePane} from "./gamePane.js";
-import {Button} from   "./button.js";
-import {Controller} from "../controller.js";
+import React, { Component } from 'react';
+import {ReactController} from "./reactController.js";
 
-export class MainMenu extends GamePane{
-    constructor(){
-        super();
-        this.addChild(this.howToPlayButton());
-        this.addChild(this.fightButton());
-        this.addChild(this.teamBuilderButton());
-        this.addChild(this.infoButton());
-        //add help button
+export class MainMenu extends Component {
+    howToPlay(){
+        console.log("How to play not created yet");
+    }
+    fight(){
+        //make sure user is logged in
+        //make sure they have at least 2 teams
+        this.props.controller.setView(ReactController.TEAM_SELECT);
+    }
+    teamBuilder(){
+        //make sure user is logged in
+        console.log("build teams");
+    }
+    info(){
+        console.log("Info not created yet");
     }
     
-    howToPlayButton(){
-        let ret = new Button("How to play");
-        ret.setPos(0, 0);
-        ret.setSize(25, 100);
-        ret.setColor("green");
-        ret.addOnClick(()=>{
-            console.log("How to play not created yet");
-        });
-        return ret;
-    }
-    
-    fightButton(){
-        let ret = new Button("Fight!");
-        ret.setPos(25, 0);
-        ret.setSize(25, 100);
-        ret.setColor("red");
-        ret.addOnClick(()=>{
-            this.controller.setView(Controller.TEAM_SELECT);
-        });
-        return ret;
-    }
-    
-    teamBuilderButton(){
-        let ret = new Button("Teambuilder");
-        ret.setPos(50, 0);
-        ret.setSize(25, 100);
-        ret.setColor("blue");
-        ret.addOnClick(()=>{
-            this.controller.setView(Controller.TEAM_BUILDER);
-        });
-        return ret;
-    }
-    
-    infoButton(){
-        let ret = new Button("Info");
-        ret.setPos(75, 0);
-        ret.setSize(25, 100);
-        ret.setColor("yellow");
-        ret.addOnClick(()=>{
-            console.log("Info not created yet");
-        });
-        return ret;
+    render(){
+        return (
+            <div className="MainMenu">
+                <div className="MainMenuItem" id="HowToPlayButton" onClick={this.howToPlay.bind(this)}>
+                    <p> How to Play </p>
+                </div>
+                <div className="MainMenuItem" id="FightButton" onClick={this.fight.bind(this)}>
+                    <p> Fight! </p>
+                </div>
+                <div className="MainMenuItem" id="TeamBuilderButton" onClick={this.teamBuilder.bind(this)}>
+                    <p> Teambuilder </p>
+                </div>
+                <div className="MainMenuItem" id="InfoButton" onClick={this.info.bind(this)}>
+                    <p> Info </p>
+                </div>
+            </div>
+        );
     }
 }
